cluster: refactor shared waiter_queue into utils
This code was added concurrently for very similar use
cases on different branches.
