+++ src/v/cluster/feature_table.cc
@@ -92,27 +92,7 @@ void feature_table::set_active_version(cluster_version v) {
     _active_features_mask = 0x0;
     for (const auto& f : get_active_features()) {
         _active_features_mask |= uint64_t(f);
-    }
-
-    // Check all pending waiters and kick any that are ready.  Do exhaustive
-    // pass: feature updates are rare and waiter count is small, order of the
-    // number of subsystems.
-    auto waiters = std::exchange(_waiters, {});
-    for (auto& wi : waiters) {
-        if (is_active(wi->f)) {
-            vlog(
-              clusterlog.trace,
-              "Triggering waiter for feature {}",
-              to_string_view(wi->f));
-            wi->p.set_value();
-        } else {
-            vlog(
-              clusterlog.trace,
-              "Not ready for waiter (version={}, waiter wants feature {})",
-              _active_version,
-              to_string_view(wi->f));
-            _waiters.push_back(std::move(wi));
-        }
+        _waiters.notify(f);
     }
 }
 
@@ -127,29 +107,7 @@ ss::future<> feature_table::await_feature(feature f, ss::abort_source& as) {
         return ss::now();
     } else {
         vlog(clusterlog.trace, "Waiting for feature {}", to_string_view(f));
-        auto item = std::make_unique<wait_item>(f);
-        auto fut = item->p.get_future();
-
-        auto sub_opt = as.subscribe(
-          [item_ptr = item->weak_from_this()]() noexcept {
-              if (item_ptr) {
-                  vlog(
-                    clusterlog.trace,
-                    "Abort source fired for waiter on {}",
-                    to_string_view(item_ptr->f));
-                  item_ptr->p.set_exception(ss::abort_requested_exception());
-              }
-          });
-
-        if (!sub_opt) {
-            // Abort source already fired!
-            return ss::make_exception_future<>(ss::abort_requested_exception());
-        } else {
-            item->abort_sub = std::move(*sub_opt);
-        }
-
-        _waiters.emplace_back(std::move(item));
-        return fut;
+        return _waiters.await(f, as);
     }
 }
